split logics into files
